增加Lua

+++ src/jmsg_type.cpp
@@ -1,7 +1,7 @@
 #include "jmsg_type.h"
 #include "jmsg_field.h"
 JMsgField* JMsgType::getFieldById(int fieldId) {
-	for(int i = 0; i < m_vecFields.size(); i++) {
+	for(size_t i = 0; i < m_vecFields.size(); i++) {
 		JMsgField* field = m_vecFields[i];
 
 		if(field->m_id == fieldId) {
